skip_requirement_library.txt
aiohttp==3.8.3
aiosignal==1.2.0
appengine-python-standard==1.0.0
appnope==0.1.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
db-dtypes==1.0.4
debugpy==1.6.2
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==2.2.2
frozendict==2.3.4
frozenlist==1.3.1
fsspec==2022.8.2
gcsfs==2022.8.2
google-api-core==2.10.1
google-auth==2.12.0
google-auth-oauthlib==0.5.3
google-cloud==0.34.0
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
idna==3.3
importlib-metadata==5.0.0
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.11.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mock==4.0.3
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.23.1
oauthlib==3.2.1
packaging==21.3
pandas==1.4.3
pandas-gbq==0.17.9
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
proto-plus==1.22.1
protobuf==4.21.7
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydata-google-auth==1.4.0
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
stack-data==0.3.0
toml==0.10.2
tornado==6.2
traitlets==5.3.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0
